fix(Promise): Result marshalling
Fixed PromiseOfT result marshalling for object types by casting the
handle value from `int` to `IntPtr` first.

Also added regression tests for promises of `JsObject`-derived types.
